Remove CFP, Talk Ideas language
Since the CFP is now closed, we need a new version where CFP submission language is removed, and talk ideas is no longer present.  This means removing CFP and Talk Ideas from the jumbotron, and augmenting the language for the index to thank folks who submitted talks (and to check back later for more information about selected talks).