Add No Starch Press as In-Kind Sponsor; misc. Cleanup
Added No Starch Press as an in-kind sponsor for the sponsors page.  This includes their img and blurb.

I also took the time to add a blurb for O'Reilly since none was present (used information from their website 'corporate info' page.)

Finally, I noticed that we had duplication in thanking the sponsors.  I removed the second instance of the thanking.