Add schedule; give basic information about speakers
In order fully add the schedule, I needed to fill in basic information
about the speakers that have not yet confirmed. So, a huge caution about
pushing this branch - we absolutely need to confirm with all the speakers
before doing a merge and middleman deploy of this code.

That said, HOORAY!  We have a schedule!